requirements.txt
torch>=2.0,<3.0
torchvision>=0.15,<1.0
Pillow>=10.0,<12.0
boto3>=1.28,<2.0
requests>=2.31,<3.0
